@@ -1,6 +1,6 @@
 {
   "name": "postcss-load-config",
-  "version": "4.0.2",
+  "version": "6.0.1",
   "description": "Autoload Config for PostCSS",
   "main": "src/index.js",
   "types": "src/index.d.ts",
@@ -8,7 +8,7 @@
     "src"
   ],
   "engines": {
-    "node": ">= 14"
+    "node": ">= 18"
   },
   "funding": [
     {
@@ -21,18 +21,25 @@
     }
   ],
   "dependencies": {
-    "lilconfig": "^3.0.0",
-    "yaml": "^2.3.4"
+    "lilconfig": "^3.1.1"
   },
   "peerDependencies": {
+    "jiti": ">=1.21.0",
     "postcss": ">=8.0.9",
-    "ts-node": ">=9.0.0"
+    "tsx": "^4.8.1",
+    "yaml": "^2.4.2"
   },
   "peerDependenciesMeta": {
-    "ts-node": {
+    "jiti": {
       "optional": true
     },
     "postcss": {
+      "optional": true
+    },
+    "tsx": {
+      "optional": true
+    },
+    "yaml": {
       "optional": true
     }
   },
