@@ -1,6 +1,6 @@
 {
   "name": "tailwindcss",
-  "version": "3.4.13",
+  "version": "3.4.18",
   "description": "A utility-first CSS framework for rapidly building custom user interfaces.",
   "license": "MIT",
   "main": "lib/index.js",
@@ -43,49 +43,49 @@
     "*.js"
   ],
   "devDependencies": {
-    "@swc/cli": "^0.1.62",
-    "@swc/core": "^1.3.55",
-    "@swc/jest": "^0.2.26",
-    "@swc/register": "^0.1.10",
-    "autoprefixer": "^10.4.14",
-    "browserslist": "^4.21.5",
-    "concurrently": "^8.0.1",
+    "@swc/cli": "0.1.62",
+    "@swc/core": "1.3.55",
+    "@swc/jest": "0.2.26",
+    "@swc/register": "0.1.10",
+    "autoprefixer": "^10.4.20",
+    "browserslist": "^4.24.0",
+    "concurrently": "^8.2.2",
     "cssnano": "^6.1.2",
-    "esbuild": "^0.20.2",
-    "eslint": "^8.39.0",
-    "eslint-config-prettier": "^8.8.0",
+    "esbuild": "^0.24.0",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^29.6.0",
-    "jest-diff": "^29.6.0",
-    "lightningcss": "1.24.1",
+    "jest": "^29.7.0",
+    "jest-diff": "^29.7.0",
+    "lightningcss": "1.27.0",
     "prettier": "^2.8.8",
-    "rimraf": "^5.0.0",
-    "source-map-js": "^1.0.2",
-    "turbo": "^1.9.3"
+    "rimraf": "^5.0.10",
+    "source-map-js": "^1.2.1",
+    "turbo": "^1.13.4"
   },
   "dependencies": {
     "@alloc/quick-lru": "^5.2.0",
     "arg": "^5.0.2",
-    "chokidar": "^3.5.3",
+    "chokidar": "^3.6.0",
     "didyoumean": "^1.2.2",
     "dlv": "^1.1.3",
-    "fast-glob": "^3.3.0",
+    "fast-glob": "^3.3.2",
     "glob-parent": "^6.0.2",
     "is-glob": "^4.0.3",
-    "jiti": "^1.21.0",
-    "lilconfig": "^2.1.0",
-    "micromatch": "^4.0.5",
+    "jiti": "^1.21.7",
+    "lilconfig": "^3.1.3",
+    "micromatch": "^4.0.8",
     "normalize-path": "^3.0.0",
     "object-hash": "^3.0.0",
-    "picocolors": "^1.0.0",
-    "postcss": "^8.4.23",
+    "picocolors": "^1.1.1",
+    "postcss": "^8.4.47",
     "postcss-import": "^15.1.0",
     "postcss-js": "^4.0.1",
-    "postcss-load-config": "^4.0.1",
-    "postcss-nested": "^6.0.1",
-    "postcss-selector-parser": "^6.0.11",
-    "resolve": "^1.22.2",
-    "sucrase": "^3.32.0"
+    "postcss-load-config": "^4.0.2 || ^5.0 || ^6.0",
+    "postcss-nested": "^6.2.0",
+    "postcss-selector-parser": "^6.1.2",
+    "resolve": "^1.22.8",
+    "sucrase": "^3.35.0"
   },
   "browserslist": [
     "> 1%",
